django-backend/requirements.txt
asgiref==3.8.1
black==23.3.0
build==1.2.2.post1
cached-property==2.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Django==4.2.2
graphene-django==3.2.3
django-cors-headers==4.7.0
django-dbcleanup==0.1.4
django-filter==25.1
django-fixture-magic==0.1.5
django-url-filter==0.3.15
djangorestframework==3.16.0
docopt==0.6.2
enum-compat==0.0.3
future==1.0.0
idna==3.10
mypy_extensions==1.1.0
narwhals==1.38.2
numpy==2.2.5
packaging==25.0
platformdirs==4.3.8
plotly==6.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
yarg==0.1.10
